+++ CHANGELOG.md
@@ -9,10 +9,16 @@
 
 ### Changed
 
+- Update `innmind/immutable` to version `~4.0`
 - `Innmind\Doctrine\Repository::matching()` now returns an instance of `Innmind\Doctrine\Matching`
 - `Innmind\Doctrine\Repository::all()` now returns an instance of `Innmind\Doctrine\Matching`
 - Update `innmind/specification` to version `~3.0`
 
 ### Removed
 
+- `Innmind\Doctrine\Sequence`
+- `Innmind\Doctrine\Sequence\Concrete`
+- `Innmind\Doctrine\Sequence\DeferFindBy`
+- `Innmind\Doctrine\Sequence\DeferQuery`
+- `Innmind\Doctrine\unwrap`
 - Support for php `7.4` and `8.0`